// src/components/pages/Report/ReportBmppGroupDpop/DetailPage/components/Detail/Detail.types.ts
export type DetailProps = {
  data?: {
    groupId?: string;
    groupName?: string;
    groupType?: string;
    groupRelation?: string;
    division?: string;
    modal?: string;
    totalFasilitasExistingGroup?: string;
    kelonggaranBmppGroup?: string;
    persyaratanBmppGroup?: string;
    presentaseRealisasiBmppGroup?: string;
    kesimpulan?: string;
    informasiPenarikanDataModalPer?: string;
    informasiTanggalInputDataModal?: string;
    informasiFasilitasExisting?: string;
    createdDate?: string;
  };
  isLoading: boolean;
};